API_Creation_Backend/requirements.txt


fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
aiomysql==0.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
